Clear app.py to move components function to components.py
